+++ packages/smooth_app/lib/cards/product_cards/product_list_preview_helper.dart
@@ -6,7 +6,7 @@ import 'package:openfoodfacts/model/Product.dart';
 import 'package:smooth_ui_library/widgets/smooth_product_image.dart';
 
 // Project imports:
-import 'package:smooth_app/pages/product_page.dart';
+import 'package:smooth_app/pages/product/product_page.dart';
 
 class ProductListPreviewHelper extends StatelessWidget {
   const ProductListPreviewHelper({
