+++ packages/smooth_app/lib/cards/product_cards/smooth_product_card_found.dart
@@ -14,7 +14,7 @@ import 'package:smooth_ui_library/widgets/smooth_product_image.dart';
 // Project imports:
 import 'package:smooth_app/cards/data_cards/attribute_chip.dart';
 import 'package:smooth_app/data_models/user_preferences_model.dart';
-import 'package:smooth_app/pages/product_page.dart';
+import 'package:smooth_app/pages/product/product_page.dart';
 import 'package:smooth_app/temp/user_preferences.dart';
 
 class SmoothProductCardFound extends StatelessWidget {
