+++ packages/smooth_app/lib/cards/product_cards/smooth_product_card_not_found.dart
@@ -6,7 +6,7 @@ import 'package:openfoodfacts/model/Product.dart';
 import 'package:smooth_ui_library/buttons/smooth_simple_button.dart';
 
 // Project imports:
-import 'package:smooth_app/pages/product_page.dart';
+import 'package:smooth_app/pages/product/product_page.dart';
 
 class SmoothProductCardNotFound extends StatelessWidget {
   const SmoothProductCardNotFound({
