+++ packages/smooth_app/lib/pages/choose_page.dart
@@ -18,9 +18,9 @@ import 'package:smooth_app/database/group_product_query.dart';
 import 'package:smooth_app/database/keywords_product_query.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/product_query.dart';
-import 'package:smooth_app/pages/product_dialog_helper.dart';
-import 'package:smooth_app/pages/product_page.dart';
-import 'package:smooth_app/pages/product_query_page_helper.dart';
+import 'package:smooth_app/pages/product/common/product_dialog_helper.dart';
+import 'package:smooth_app/pages/product/product_page.dart';
+import 'package:smooth_app/pages/product/common/product_query_page_helper.dart';
 
 class ChoosePage extends StatefulWidget {
   @override
