+++ packages/smooth_app/lib/pages/list_page.dart
@@ -10,7 +10,7 @@ import 'package:smooth_app/cards/product_cards/product_list_preview.dart';
 import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/database/dao_product_list.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/pages/product_list_dialog_helper.dart';
+import 'package:smooth_app/pages/product/common/product_list_dialog_helper.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 
 class ListPage extends StatefulWidget {
