+++ packages/smooth_app/lib/pages/pantry/common/pantry_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 // Project imports:
 import 'package:smooth_app/data_models/pantry.dart';
-import 'package:smooth_app/pages/pantry_page.dart';
+import 'package:smooth_app/pages/pantry/pantry_page.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 
 /// A button for a pantry, with the corresponding color, icon, name and shape
