+++ packages/smooth_app/lib/pages/pantry/common/pantry_dialog_helper.dart
@@ -3,39 +3,51 @@ import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:smooth_app/temp/user_preferences.dart';
 import 'package:smooth_ui_library/buttons/smooth_simple_button.dart';
 import 'package:smooth_ui_library/dialogs/smooth_alert_dialog.dart';
 
 // Project imports:
-import 'package:smooth_app/data_models/product_list.dart';
-import 'package:smooth_app/database/dao_product_list.dart';
-import 'package:smooth_app/pages/product_list_page.dart';
-import 'package:smooth_app/pages/product_query_page_helper.dart';
+import 'package:smooth_app/data_models/pantry.dart';
+import 'package:smooth_app/pages/pantry/pantry_page.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 
-class ProductListDialogHelper {
-  static const String _TRANSLATE_ME_WANT_TO_DELETE =
-      'Do you want to delete this product list?';
-  static const String _TRANSLATE_ME_NEW_LIST = 'New list';
-  static const String _TRANSLATE_ME_RENAME_LIST = 'Rename list';
+/// A dialog helper for pantries
+class PantryDialogHelper {
+  static const String _TRANSLATE_ME_WANT_TO_DELETE_PANTRY =
+      'Do you want to delete this pantry?';
+  static const String _TRANSLATE_ME_WANT_TO_DELETE_SHOPPING =
+      'Do you want to delete this shopping list?';
+  static const String _TRANSLATE_ME_NEW_LIST_PANTRY = 'New pantry';
+  static const String _TRANSLATE_ME_NEW_LIST_SHOPPING = 'New shopping list';
+  static const String _TRANSLATE_ME_RENAME_LIST_PANTRY = 'Rename pantry';
+  static const String _TRANSLATE_ME_RENAME_LIST_SHOPPING =
+      'Rename shopping list';
   static const String _TRANSLATE_ME_CHANGE_ICON = 'Change icon';
-  static const String _TRANSLATE_ME_HINT = 'My custom list';
+  static const String _TRANSLATE_ME_HINT_PANTRY = 'My own pantry';
+  static const String _TRANSLATE_ME_HINT_SHOPPING = 'My shopping list';
   static const String _TRANSLATE_ME_EMPTY = 'Please enter some text';
-  static const String _TRANSLATE_ME_ALREADY_OTHER =
-      'There\'s already a list with that name';
+  static const String _TRANSLATE_ME_ALREADY_OTHER_PANTRY =
+      'There\'s already a pantry with that name';
+  static const String _TRANSLATE_ME_ALREADY_OTHER_SHOPPING =
+      'There\'s already a shopping list with that name';
   static const String _TRANSLATE_ME_ALREADY_SAME = 'That\'s the same name!';
   static const String _TRANSLATE_ME_CANCEL = 'Cancel';
 
   static Future<bool> openDelete(
     final BuildContext context,
-    final DaoProductList daoProductList,
-    final ProductList productList,
+    final List<Pantry> pantries,
+    final int index,
   ) async =>
       await showDialog<bool>(
         context: context,
         builder: (BuildContext context) => SmoothAlertDialog(
           close: false,
-          body: const Text(_TRANSLATE_ME_WANT_TO_DELETE),
+          body: Text(
+            pantries[index].pantryType == PantryType.PANTRY
+                ? _TRANSLATE_ME_WANT_TO_DELETE_PANTRY
+                : _TRANSLATE_ME_WANT_TO_DELETE_SHOPPING,
+          ),
           actions: <SmoothSimpleButton>[
             SmoothSimpleButton(
               text: AppLocalizations.of(context).no,
@@ -46,7 +58,7 @@ class ProductListDialogHelper {
               text: AppLocalizations.of(context).yes,
               important: true,
               onPressed: () async {
-                await daoProductList.delete(productList);
+                pantries.removeAt(index);
                 Navigator.pop(context, true);
               },
             ),
@@ -54,43 +66,53 @@ class ProductListDialogHelper {
         ),
       );
 
-  static Future<ProductList> openNew(
+  static Future<String> openNew(
     final BuildContext context,
-    final DaoProductList daoProductList,
-    final List<ProductList> list,
+    final List<Pantry> pantries,
+    final PantryType pantryType,
+    final UserPreferences userPreferences,
   ) async {
+    String newPantryName;
     final GlobalKey<FormState> formKey = GlobalKey<FormState>();
-    ProductList newProductList;
-    return await showDialog<ProductList>(
+    return await showDialog<String>(
       context: context,
       builder: (BuildContext context) => SmoothAlertDialog(
         close: false,
-        title: _TRANSLATE_ME_NEW_LIST,
+        title: pantryType == PantryType.PANTRY
+            ? _TRANSLATE_ME_NEW_LIST_PANTRY
+            : _TRANSLATE_ME_NEW_LIST_SHOPPING,
         body: Form(
           key: formKey,
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               TextFormField(
-                decoration: const InputDecoration(
-                  hintText: _TRANSLATE_ME_HINT,
+                decoration: InputDecoration(
+                  hintText: pantryType == PantryType.PANTRY
+                      ? _TRANSLATE_ME_HINT_PANTRY
+                      : _TRANSLATE_ME_HINT_SHOPPING,
                 ),
                 validator: (final String value) {
                   if (value.isEmpty) {
                     return _TRANSLATE_ME_EMPTY;
                   }
-                  if (list == null) {
+                  if (pantries == null) {
                     return null;
                   }
-                  newProductList = ProductList(
-                    listType: ProductList.LIST_TYPE_USER_DEFINED,
-                    parameters: value,
-                  );
-                  for (final ProductList productList in list) {
-                    if (productList.lousyKey == newProductList.lousyKey) {
-                      return _TRANSLATE_ME_ALREADY_OTHER;
+                  for (int i = 0; i < pantries.length; i++) {
+                    if (value == pantries[i].name) {
+                      return pantryType == PantryType.PANTRY
+                          ? _TRANSLATE_ME_ALREADY_OTHER_PANTRY
+                          : _TRANSLATE_ME_ALREADY_OTHER_SHOPPING;
                     }
                   }
+                  pantries.add(Pantry(name: value, pantryType: pantryType));
+                  Pantry.putAll(
+                    userPreferences,
+                    pantries,
+                    pantryType,
+                  );
+                  newPantryName = value;
                   return null;
                 },
               ),
@@ -109,23 +131,25 @@ class ProductListDialogHelper {
               if (!formKey.currentState.validate()) {
                 return;
               }
-              if (await daoProductList.get(newProductList)) {
-                // TODO(monsieurtanuki): unexpected, but do something!
-                return;
-              }
-              await daoProductList.put(newProductList);
-              Navigator.pop(context, newProductList);
-              await Navigator.push<dynamic>(
-                context,
-                MaterialPageRoute<dynamic>(
-                  builder: (BuildContext context) => ProductListPage(
-                    newProductList,
-                    reverse: ProductQueryPageHelper.isListReversed(
-                      newProductList,
+              Navigator.pop(context, newPantryName);
+
+              int index = 0;
+              for (final Pantry pantry in pantries) {
+                if (pantry.name == newPantryName) {
+                  await Navigator.push<dynamic>(
+                    context,
+                    MaterialPageRoute<dynamic>(
+                      builder: (BuildContext context) => PantryPage(
+                        pantries,
+                        index,
+                        pantryType,
+                      ),
                     ),
-                  ),
-                ),
-              );
+                  );
+                  return;
+                }
+                index++;
+              }
             },
             important: true,
           ),
@@ -134,46 +158,49 @@ class ProductListDialogHelper {
     );
   }
 
-  static Future<ProductList> openRename(
+  static Future<bool> openRename(
     final BuildContext context,
-    final DaoProductList daoProductList,
-    final ProductList productList,
+    final List<Pantry> pantries,
+    final int index,
   ) async {
     final GlobalKey<FormState> formKey = GlobalKey<FormState>();
-    final List<ProductList> list =
-        await daoProductList.getAll(withStats: false);
-    ProductList newProductList;
-    return await showDialog<ProductList>(
+    return await showDialog<bool>(
       context: context,
       builder: (BuildContext context) => SmoothAlertDialog(
         close: false,
-        title: _TRANSLATE_ME_RENAME_LIST,
+        title: pantries[index].pantryType == PantryType.PANTRY
+            ? _TRANSLATE_ME_RENAME_LIST_PANTRY
+            : _TRANSLATE_ME_RENAME_LIST_SHOPPING,
         body: Form(
           key: formKey,
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               TextFormField(
-                initialValue: productList.parameters,
-                decoration: const InputDecoration(
-                  hintText: _TRANSLATE_ME_HINT,
+                initialValue: pantries[index].name,
+                decoration: InputDecoration(
+                  hintText: pantries[index].pantryType == PantryType.PANTRY
+                      ? _TRANSLATE_ME_HINT_PANTRY
+                      : _TRANSLATE_ME_HINT_SHOPPING,
                 ),
                 validator: (final String value) {
                   if (value.isEmpty) {
                     return _TRANSLATE_ME_EMPTY;
                   }
-                  newProductList = ProductList(
-                    listType: ProductList.LIST_TYPE_USER_DEFINED,
-                    parameters: value,
-                  )..extraTags = productList.extraTags;
-                  for (final ProductList item in list) {
-                    if (item.lousyKey == newProductList.lousyKey) {
-                      if (item.lousyKey == productList.lousyKey) {
+                  if (pantries == null) {
+                    return null;
+                  }
+                  for (int i = 0; i < pantries.length; i++) {
+                    if (value == pantries[i].name) {
+                      if (i == index) {
                         return _TRANSLATE_ME_ALREADY_SAME;
                       }
-                      return _TRANSLATE_ME_ALREADY_OTHER;
+                      return pantries[index].pantryType == PantryType.PANTRY
+                          ? _TRANSLATE_ME_ALREADY_OTHER_PANTRY
+                          : _TRANSLATE_ME_ALREADY_OTHER_SHOPPING;
                     }
                   }
+                  pantries[index].name = value;
                   return null;
                 },
               ),
@@ -183,7 +210,7 @@ class ProductListDialogHelper {
         actions: <SmoothSimpleButton>[
           SmoothSimpleButton(
             text: _TRANSLATE_ME_CANCEL,
-            onPressed: () => Navigator.pop(context, null),
+            onPressed: () => Navigator.pop(context, false),
             important: false,
           ),
           SmoothSimpleButton(
@@ -192,13 +219,7 @@ class ProductListDialogHelper {
               if (!formKey.currentState.validate()) {
                 return;
               }
-              if (!await daoProductList.rename(
-                  productList, newProductList.parameters)) {
-                // TODO(monsieurtanuki): unexpected, but do something!
-                return;
-              }
-              await daoProductList.get(newProductList);
-              Navigator.pop(context, newProductList);
+              Navigator.pop(context, true);
             },
             important: true,
           ),
@@ -209,10 +230,12 @@ class ProductListDialogHelper {
 
   static Future<bool> openChangeIcon(
     final BuildContext context,
-    final DaoProductList daoProductList,
-    final ProductList productList,
+    final List<Pantry> pantries,
+    final int index,
   ) async {
-    final List<String> orderedIcons = productList.getPossibleIcons();
+    final Pantry pantry = pantries[index];
+    final List<String> orderedIcons = pantries[index].getPossibleIcons();
+    const List<String> orderedColors = Pantry.ORDERED_COLORS;
     final double size = MediaQuery.of(context).size.width / 8;
     return await showDialog<bool>(
       context: context,
@@ -220,29 +243,28 @@ class ProductListDialogHelper {
         close: false,
         title: _TRANSLATE_ME_CHANGE_ICON,
         body: Container(
-          width: ProductList.ORDERED_COLORS.length.toDouble() * size,
+          width: orderedColors.length.toDouble() * size,
           height: orderedIcons.length.toDouble() * size,
           child: GridView.count(
             crossAxisCount: 5,
             childAspectRatio: 1,
             children: List<Widget>.generate(
-              ProductList.ORDERED_COLORS.length * orderedIcons.length,
+              orderedColors.length * orderedIcons.length,
               (final int index) {
-                final String colorTag = ProductList
-                    .ORDERED_COLORS[index % ProductList.ORDERED_COLORS.length];
+                final String colorTag =
+                    orderedColors[index % orderedColors.length];
                 final String iconTag =
-                    orderedIcons[index ~/ ProductList.ORDERED_COLORS.length];
+                    orderedIcons[index ~/ orderedColors.length];
                 return IconButton(
-                  icon: ProductList.getReferenceIcon(
+                  icon: pantry.getReferenceIcon(
                     colorScheme: Theme.of(context).colorScheme,
                     colorTag: colorTag,
                     iconTag: iconTag,
                     colorDestination: ColorDestination.SURFACE_FOREGROUND,
                   ),
                   onPressed: () async {
-                    productList.colorTag = colorTag;
-                    productList.iconTag = iconTag;
-                    await daoProductList.put(productList);
+                    pantry.colorTag = colorTag;
+                    pantry.iconTag = iconTag;
                     Navigator.pop(context, true);
                   },
                 );
