+++ packages/smooth_app/lib/pages/product/common/product_dialog_helper.dart
