+++ packages/smooth_app/lib/pages/product/common/product_query_page_helper.dart
@@ -15,7 +15,7 @@ import 'package:smooth_app/data_models/query_product_list_supplier.dart';
 import 'package:smooth_app/database/dao_product_list.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/product_query.dart';
-import 'package:smooth_app/pages/product_query_page.dart';
+import 'package:smooth_app/pages/product/common/product_query_page.dart';
 
 class ProductQueryPageHelper {
   Future<void> openBestChoice({
