+++ packages/smooth_app/lib/pages/scan/scan_page.dart
@@ -10,7 +10,7 @@ import 'package:smooth_ui_library/widgets/smooth_toggle.dart';
 import 'package:smooth_app/data_models/continuous_scan_model.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/product_query.dart';
-import 'package:smooth_app/pages/continuous_scan_page.dart';
+import 'package:smooth_app/pages/scan/continuous_scan_page.dart';
 
 class ScanPage extends StatelessWidget {
   const ScanPage({
